requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
asgiref==3.11.0
attrs==25.4.0
Authlib==1.6.6
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
circuitbreaker==2.1.3
click==8.3.1
colorama==0.4.6
cryptography==45.0.7
Cython==3.2.4
discord.py==2.6.4
distro==1.9.0
Django==6.0.1
Faker==40.1.2
fastapi==0.128.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.57.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio @ file:///home/runner/.termux-build/python-grpcio/src
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.10.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
numpy @ file:///home/runner/.termux-build/python-numpy/src/dist/numpy-2.2.5-cp312-cp312-linux_aarch64.whl#sha256=b28ce81b8c071c0f1f798027067c41c371aba2bf5b29a6649f10dd35a09ce4b1
oci==2.164.2
oci-cli==3.71.4
openai==0.28.0
pandas @ file:///home/builder/.termux-build/python-pandas/src/dist/pandas-3.0.0-cp312-cp312-linux_aarch64.whl#sha256=24c6bb72c60713808d1cb54200f52895d4474b89535dd059c96e3f72164dcde2
pillow==12.1.0
prompt-toolkit==3.0.43
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyOpenSSL==25.1.0
pyparsing==3.3.1
pypdf==6.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.7
requests==2.32.5
rich==14.2.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
tenacity==9.1.2
terminaltables==3.1.10
termuxgui==0.1.6
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
vritraai==0.30.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.5
wheel==0.45.1
yarl==1.22.0
